@@ -1,6 +1,6 @@
 module github.com/favonia/cloudflare-ddns
 
-go 1.20
+go 1.21
 
 require (
 	github.com/cloudflare/cloudflare-go v0.75.0
@@ -9,7 +9,6 @@ require (
 	github.com/jellydator/ttlcache/v3 v3.0.1
 	github.com/robfig/cron/v3 v3.0.1
 	github.com/stretchr/testify v1.8.4
-	golang.org/x/exp v0.0.0-20230321023759-10a507213a29
 	golang.org/x/net v0.14.0
 	kernel.org/pub/linux/libs/security/libcap/cap v1.2.69
 )
@@ -21,7 +20,7 @@ require (
 	github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
 	github.com/kr/text v0.2.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	golang.org/x/sync v0.1.0 // indirect
+	golang.org/x/sync v0.3.0 // indirect
 	golang.org/x/text v0.12.0 // indirect
 	golang.org/x/time v0.3.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
